refactor: move into common
