+++ libs/common/src/platform/ipc/ipc-message.ts
@@ -1,0 +1,8 @@
+export interface IpcMessage {
+  type: "bitwarden-ipc-message";
+  payload: Uint8Array;
+}
+
+export function isIpcMessage(message: any): message is IpcMessage {
+  return message.type === "bitwarden-ipc-message";
+}
