+++ libs/common/src/platform/ipc/ipc.service.ts
@@ -1,7 +1,11 @@
 // import { Manager } from "@bitwarden/sdk-internal";
 
+import { SdkService } from "../abstractions/sdk/sdk.service";
+
 export abstract class IpcService {
   // protected manager = new Manager();
+
+  constructor(private sdkService: SdkService) {}
 
   async init() {}
 }
