+++ libs/platform/src/index.ts
@@ -1,2 +1,1 @@
 export * from "./services/browser-service";
-export * from "./ipc";
