--- libs/platform/src/ipc/ipc-message.ts
@@ -1,3 +1,0 @@
-export interface IpcMessage {
-  payload: Uint8Array;
-}
